+++ src/pydocstyle/utils.py
@@ -1,5 +1,4 @@
 """General shared utilities."""
-import ast
 import logging
 import re
 from itertools import tee, zip_longest